adjust byte to support same range as max page size

+++ utility/position.go
@@ -3,5 +3,5 @@ package utility
 // Position represents a position in the file
 type Position struct {
 	Page uint32
-	Byte uint8
+	Byte uint16
 }
